Refactor post_save handlers
