+++ addons/googledrive/models.py
@@ -3,17 +3,12 @@
 """
 import os
 
-from django.db.models.signals import post_save
-from django.dispatch import receiver
-
 from addons.base.models import (BaseOAuthNodeSettings, BaseOAuthUserSettings,
                                 BaseStorageAddon)
 from django.db import models
 
-from addons.iqbrims.utils import copy_node_auth
 from framework.auth import Auth
 from framework.exceptions import HTTPError
-from osf.models import RdmAddonOption
 from osf.models.external import ExternalProvider
 from osf.models.files import File, Folder, BaseFileNode
 from addons.base import exceptions
@@ -22,9 +17,7 @@ from addons.googledrive.client import (GoogleAuthClient,
                                                GoogleDriveClient)
 from addons.googledrive.serializer import GoogleDriveSerializer
 from addons.googledrive.utils import to_hgrid
-from website import settings
 from website.util import api_v2_url
-from addons.iqbrims.apps import IQBRIMSAddonConfig
 
 # from website.files.models.ext import PathFollowingFileNode
 
@@ -250,19 +243,3 @@ class NodeSettings(BaseOAuthNodeSettings, BaseStorageAddon):
     def on_delete(self):
         self.deauthorize(add_log=False)
         self.save()
-
-
-@receiver(post_save, sender=NodeSettings)
-def follow_googledrive_node_settings_to_iqbrims(sender, instance, created, **kwargs):
-    if IQBRIMSAddonConfig.short_name not in settings.ADDONS_AVAILABLE_DICT:
-        return
-
-    node = instance.owner
-    is_management_node = RdmAddonOption.objects.filter(
-        provider=IQBRIMSAddonConfig.short_name,
-        management_node=node,
-        is_allowed=True
-    ).exists()
-
-    if is_management_node and node.has_addon(IQBRIMSAddonConfig.short_name):
-        copy_node_auth(node, instance)
