+++ osf/models/node.py
@@ -7,7 +7,7 @@ import warnings
 import httplib
 
 import bson
-from django.db.models import Q, Subquery
+from django.db.models import Q
 from dirtyfields import DirtyFieldsMixin
 from django.apps import apps
 from django_bulk_update.helper import bulk_update
@@ -30,7 +30,6 @@ from guardian.models import (
 )
 from guardian.shortcuts import get_objects_for_user, get_groups_with_perms, get_group_perms
 
-from addons.iqbrims.apps import IQBRIMSAddonConfig
 from framework import status
 from framework.auth import oauth_scopes
 from framework.celery_tasks.handlers import enqueue_task, get_task_from_queue
@@ -2518,26 +2517,3 @@ def set_parent_and_root(sender, instance, created, *args, **kwargs):
     if not instance.root:
         instance.root = instance.get_root()
         instance.save()
-
-
-@receiver(post_save, sender=Node)
-def add_iqbrims_addon(sender, instance, created, **kwargs):
-    from osf.models import RdmAddonOption
-
-    if IQBRIMSAddonConfig.short_name not in settings.ADDONS_AVAILABLE_DICT:
-        return
-
-    inst_ids = instance.affiliated_institutions.values('id')
-    addon_option = RdmAddonOption.objects.filter(
-        provider=IQBRIMSAddonConfig.short_name,
-        institution_id__in=Subquery(inst_ids),
-        management_node__isnull=False,
-        is_allowed=True
-    ).first()
-    if addon_option is None:
-        return
-    if addon_option.organizational_node is not None and \
-            not addon_option.organizational_node.is_contributor(instance.creator):
-        return
-
-    instance.add_addon(IQBRIMSAddonConfig.short_name, auth=None, log=False)
